+++ packages/ui/src/components/toaster.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useToast } from "@mcw/ui";
+import { useToast } from "../hooks/use-toast";
 import {
   Toast,
   ToastClose,
@@ -8,7 +8,7 @@ import {
   ToastProvider,
   ToastTitle,
   ToastViewport,
-} from "@mcw/ui";
+} from "./toast";
 
 export function Toaster() {
   const { toasts } = useToast();
