+++ packages/ui/src/index.ts
@@ -46,10 +46,10 @@ export * from "./components/separator";
 export * from "./components/sidebar";
 export * from "./components/skeleton";
 export * from "./components/slider";
-export * from "./components/sonner";
-export * from "./components/switch";
-export { useToast } from "./hooks/use-toast";
+export * from "./hooks/use-toast";
 export * from "./components/toast";
+export * from "./components/toaster";
+export * from "./components/switch";
 export * from "./components/time-picker";
 export * from "./components/toggle";
 export * from "./components/toggle-group";
